+++ frontend/src/constants/Routes.ts
@@ -16,9 +16,9 @@ export const SET_PASSWORD_PAGE = "/:objectID/set-password";
 
 export const CREATE_MEAL_REQUEST_PAGE = "/request/scheduling";
 
-export const SETTINGS_PAGE = "/settings";
-
 export const DASHBOARD_PAGE = "/dashboard";
+
+export const SETTINGS_PAGE = "/settings";
 
 export const EDIT_TEAM_PAGE = "/dashboard/edit-team";
